+++ src/AnnotatedRoutes/tests/App/App.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Tests\Router\App;
@@ -21,7 +14,7 @@ class App extends Kernel
 {
     protected const LOAD = [
         DiactorosBootloader::class,
-        AnnotatedRoutesBootloader::class
+        AnnotatedRoutesBootloader::class,
     ];
 
     public function getHttp(): Http