+++ src/Exceptions/tests/Renderer/PlainRendererTest.php
@@ -9,18 +9,18 @@
 
 declare(strict_types=1);
 
-namespace Spiral\Tests\Exceptions;
+namespace Spiral\Tests\Exceptions\Renderer;
 
 use Exception;
 use PHPUnit\Framework\TestCase;
-use Spiral\Exceptions\PlainHandler;
+use Spiral\Exceptions\Renderer\PlainRenderer;
 
-class PlainHandlerTest extends TestCase
+class PlainRendererTest extends TestCase
 {
     public function testRenderException(): void
     {
-        $plainHandler = new PlainHandler();
-        $result = $plainHandler->renderException(new Exception('Undefined variable $undefined'));
+        $plainHandler = new PlainRenderer();
+        $result = $plainHandler->render(new Exception('Undefined variable $undefined'));
 
         $this->assertStringContainsString(
             sprintf('Undefined variable $undefined in %s', __FILE__),