+++ src/Queue/src/Failed/LogFailedJobHandler.php
@@ -4,17 +4,17 @@ declare(strict_types=1);
 
 namespace Spiral\Queue\Failed;
 
-use Spiral\Snapshots\SnapshotterInterface;
+use Spiral\Exceptions\ExceptionReporterInterface;
 
 final class LogFailedJobHandler implements FailedJobHandlerInterface
 {
     public function __construct(
-        private readonly SnapshotterInterface $snapshotter
+        private readonly ExceptionReporterInterface $reporter
     ) {
     }
 
     public function handle(string $driver, string $queue, string $job, array $payload, \Throwable $e): void
     {
-        $this->snapshotter->register($e);
+        $this->reporter->report($e);
     }
 }